+++ CHANGELOG.md
@@ -5,6 +5,27 @@
 ## API Changes
 
 ## Bug Fixes
+* Fixed bug in able actions which could result in action buttons being in the incorrect (enabled) state if the table
+  isloaded via AJAX #1049.
+* Fixed a bug in modalShim which resulted in accesskeys being stripped from inside WDialog content #1051.
+
+## Enhancements
+
+# Release 1.3.2
+
+## Bug Fixes
+* Fixed a bug in XSLT which caused processor errors if an instance of Input has more than one validation error #1042.
+* Fixed a bug in subordinate JavaScript which caused WMultiSelectPair to fail as a subordinate trigger #1044.
+
+## Enhancements
+* Reduced logging "noise" when the corrupt characters flag is enabled #1039
+
+# Release 1.3.1
+
+## API Changes
+* Deprecated xslt server side property. Will always return true.
+
+## Bug Fixes
 
 ## Enhancements